requirements.txt
selenium==4.22.0
pytest==8.2.2
allure-pytest==2.13.5